+++ includes/content-gate/class-content-gate.php
@@ -168,6 +168,11 @@ class Content_Gate {
 	 * Register post type for custom gate.
 	 */
 	public static function register_post_type() {
+		$icon = sprintf(
+			'data:image/svg+xml;base64,%s',
+			base64_encode( Newspack_UI_Icons::get_svg( 'key' ) )
+		);
+
 		\register_post_type(
 			self::GATE_CPT,
 			[
@@ -184,6 +189,7 @@ class Content_Gate {
 				'show_ui'      => true,
 				'show_in_menu' => true,
 				'show_in_rest' => true,
+				'menu_icon'    => $icon,
 				'supports'     => [ 'editor', 'custom-fields', 'revisions', 'title' ],
 				'taxonomies'   => [ 'category', 'post_tag' ],
 			]